Removing/Compacting/Relocating some of the tests
